change the structure and add the right server.js to root
